Set immediate flag to false
Using immediate set to true can lead to BUGs and oopses when
downgrading, reverting or applying replace_all patches. There is no way
how to find out if there is a process in the old code which is being
removed. The module would be put, removed and the process will crash.

The consistency model guarantees that there is no one in the old code
when the finalization ends. Thus use it for all case to be safe.

Signed-off-by: Miroslav Benes <mbenes@suse.cz>