scripts: create kgr_patch_main.c dynamically
The kgraft-patches repository has got many branches, each corresponding
to a supported codestream. Each of those carries a potentially different
set of live (sub)patches which are controlled through the entry points in
kgr_patch_main.c. According to Miroslav, merging of a new (sub)patch
based on the pristine master is a pita due to conflicts.

Since all (sub)patches stick to certain conventions already, the required
modifications of the merging-hotspot kgr_patch_main.c are quite mechanic.
Let a script do the work.

Namely,
- insert some special @@-embraced placeholders at the few places depending
  on the actual set of (sub)patches,
- let register-patches.sh discover the available (sub)patches by searching
  for directories
- and let register-patches.sh replace those placeholders in
  kgr_patch_main.c
Finally, add a register-patches.sh invocation to tar-up.sh.

This procedure requires that a SUBPATCH located in directory SUBPATCH/
adheres to the following conventions:
- It must provide a provide a SUBPATCH/kgr_patch_SUBPATCH.h header.
- This header must provide declarations for kgr_patch_SUBPATCH_init()
  and kgr_patch_SUBPATCH_cleanup().
- This header must also #define a KGR_PATCH_SUBPATCH_FUNCS macro.
  It should expand to a comma separated list of KGR_PATCH*() entries,
  each corresponding to a function the subpatch wants to replace.

Signed-off-by: Nicolai Stange <nstange@suse.de>
[mbenes: fixed typos, empty line removed]
Signed-off-by: Miroslav Benes <mbenes@suse.cz>