shadow variables: introduce upstream patch
Joe Lawrence posted the sixth version of his shadow variable patch [1]
implementing the association of additional out-of-band data members to
existing structure instances from livepatches.

Jiri Kosina has applied this to his

  git://git.kernel.org/pub/scm/linux/kernel/git/jikos/livepatching.git for-4.15/shadow-variables

tree and thus, it's queued up and close to getting merged.

The plan is to eventually backport this shadow variable support to SLE
kernels, but we also want to have it usable from KGraft modules by now.

Port the implementation to the kraft-patches module.

Namely,
 - dump shadow.c in it's current upstream state as it is after commits
     439e7271dc2b ("livepatch: introduce shadow variable API")
     5d9da759f758 ("livepatch: __klp_shadow_get_or_alloc() is local to
                    shadow.c")
     19205da6a0da ("livepatch: Small shadow variable documentation fixes")
 - add a shadow.h header and declare the newly introduced functions there
 - and incorporate the new files into the KGraft module's build system.

[1] 1504211861-19899-2-git-send-email-joe.lawrence@redhat.com
    ("[PATCH v6] livepatch: introduce shadow variable API")

Signed-off-by: Nicolai Stange <nstange@suse.de>
Signed-off-by: Miroslav Benes <mbenes@suse.cz>