scripts/register-patches.sh: allow spaces as patched_funcs.csv separators
Currently there's one single cut(1) usage which requires that (single) tabs
are used as field separators for the patched_funcs.csv.

As the rest of the code can deal with sequences of any whitespace already,
this imposes an unnecessary restriction on the format.

Substitute that cut(1) usage by a sed(1) invocation as appropriate.

Signed-off-by: Nicolai Stange <nstange@suse.de>
Signed-off-by: Miroslav Benes <mbenes@suse.cz>