Use klp-convert where provided
klp-convert tool converts undefined symbols in a live patch kernel module
to special relocation records which are resolved by the kernel. It
allows to omit kallsyms tricks.

Wire it to the spec file and let tar-up.sh script decide if it is to be
used depending on a codestream. SLE15-SP1 is supported currently.

Signed-off-by: Miroslav Benes <mbenes@suse.cz>
Reviewed-by: Nicolai Stange <nstange@suse.de>