Define macros to switch easily between klp-convert and kallsyms
Kallsyms trick does not have to be used for resolving undefined symbols
when klp-convert is available. It would be great though to share live
patches sources between both modes of operation.

Define macros to help with the task. Their definitions depend on
whether USE_KLP_CONVERT macro is defined. tar-up.sh script is
responsible to decide.

Signed-off-by: Miroslav Benes <mbenes@suse.cz>
Reviewed-by: Nicolai Stange <nstange@suse.de>