scripts: Disable use of klp-convert
klp-convert tool was introduced to improve a situation with unexported
symbols while preparing live patches. However, it is still not stable
enough and upstream still needs to decide the purpose of the tool. Given
that it is used only for uname patch and only on SLE15-SP1 it is better
to just disable it for now.

At the same, leave the infrastructure in place, because we might use it
in the future.

Signed-off-by: Miroslav Benes <mbenes@suse.cz>