klp_syscalls.h: fix syscall prototype mismatch on s390x for kernels >= 4.17
The __SYSCALL_DEFINEx(x, name, ...) macro as defined in
arch/s390/include/asm/syscall_wrapper.h declares two protoypes for
a given syscall: __s390x_sys##name() and __se_sys##name(). The former
symbol is made to be an alias to the latter and the function arguments are
of the "real" type as specified in the macro invocation whereas the
latter's argument types are transformed into longs.

Currently the KLP_SYSCALL_SYM() helper macro from our klp_syscalls.h
evaluates to the __s390x_sys##name() variant, but its expansion result is
intended to be used with KLP_SYSCALL_DECLx(), which does the transformation
of the arguments' types to longs. This results in compilation errors due to
the syscall prototype declaration from KLP_SYSCALL_DECLx() confliciting
with the one from __SYSCALL_DEFINEx(), if visible.

The current behaviour of KLP_SYSCALL_DECLx() should be retained in order
to keep it working for the compatibility stubs, i.e. with
KLP_SYSCALL_COMPAT_STUB_SYM(). So fix the issue by making KLP_SYSCALL_SYM()
to evaluate to the __se_sys##name() variant on 390x for kernel versions >=
4.17.

Signed-off-by: Nicolai Stange <nstange@suse.de>