uname_patch: Trim klp tag to fix the overflow
SLE15-SP4 introduces an option to specify a preempt model during boot.
new_utsname->version was updated to take this into account and contains
PREEMPT_DYNAMIC tag now. In the end, there is not much space left to
include our klp_tag and it overflows. Instead of removing the tag
completely, trim it so that the user can at least easily spot that a
live patch is installed on the system.

A git HEAD revision will be stored elsewhere.

While at it, make the tag const.

References: bsc#1196281
Signed-off-by: Miroslav Benes <mbenes@suse.cz>
Reviewed-by: Nicolai Stange <nstange@suse.de>
Acked-by: Libor Pechacek <lpechacek@suse.cz>