klp_trace.h: improvement for SLE15-SP6 external symbols
Because some parts of tracepoint macro (like traceiter definition) are hidden
behind the macro, so klp-build cannot even detect those symbols.

To avoid doing some heuristics detecting `traceiter` missing definition, this
patch modifies the `KLPR_DECLARE_TRACE` macro to add the KLP_RELOC_SYMBOL
to its symbols. To do this, it needs to know the module name, which SHOULD
be always defined in the same module we are trying to LP, so I added only one
parameter (module name) putting it to both LPed object name and also to
target object name parameters.

Signed-off-by: Lukas Hruska <lhruska@suse.cz>
Signed-off-by: Nicolai Stange <nstange@suse.de>