Remove the support for different flavors
There is a support for different kernel flavors from the beginning in
our spec file. Originally, there were -default and -xen flavors.
However, it is questionable. A live patch is built against a very
specific kernel binary. Different flavors of the same kernel source can
be easily different also in this respect.

Remove it then. The build process is driven by "variant" macro deriving
from a branch name. We can stick with that.

Signed-off-by: Miroslav Benes <mbenes@suse.cz>
Reviewed-by: Nicolai Stange <nstange@suse.de>