kernel-livepatch.spec: Replace kernel-syms with kernel-<flavor>-specific dependencies (bsc#1248108)
The commit ead79afe7cbfae ("kernel-livepatch.spec: Update build
dependencies for non-default flavors") broke build of livepatches
which were built with kernel-syms-rt.

The problem is that livepatch packages for already released kernels
are built in exactly the same build environment as the initial livepatch.
The BS (Build Service) installs the build environment using the given
_buildinfo-*.xml and ignores BuildRequires. But the BuildRequires are
later checked by rpmbuild tool. It would complain when new dependencies
were added.

Unfortunately, kernel-syms-rt does not exist on SLE16. This was the main
motivation for the above mentioned commit.

But the package kernel-syms is empty. Its only purpose is to add other
dependencies. Replace it by opencoding the dependencies.

Note that the kernel devel files are historically split into various
packages, kernel-<flavor>-devel, kernel-devel-<flavor>, and
even kernel-devel. But it is enough to require kernel-<flavor>-devel
because it requires the other devel files on its own. This seems
to be true back to SLE15-SP4 at minimum.

Signed-off-by: Petr Mladek <pmladek@suse.com>
Signed-off-by: Miroslav Benes <mbenes@suse.cz>